feat: multi-channel messaging, plugin system, gateway routing, browser profiles, model failover, and full-stack UI expansion
Backend — new modules:
- Multi-channel messaging: Discord, Slack, Teams, Matrix channel drivers
- Gateway routing engine with ops control and node management
- Plugin registry with SDK scaffolding and HTTP endpoints
- Browser profile manager for isolated Playwright sessions
- Model failover with automatic provider switching
- Sender verification system for inbound message authentication
- Integration sync engine for bidirectional data flow
- Custom API framework for user-defined endpoints
- Google Workspace actions (admin, drive, calendar, gmail)
- CLI entrypoint for headless/scripted operation
- Vendored russh crate for SSH connectivity

Backend — enhancements:
- Agent core: expanded tool execution, prompt builder, task routing
- Runtime: extended orchestration and lifecycle management
- HTTP layer: webhooks, tunnel improvements, gateway/node/plugin controls
- Storage: new entities, schema migrations, query optimizations
- MCP registry: dynamic tool discovery and client improvements
- Config: model failover settings, channel configs, gateway params
- Sentinel: backup scheduling, health monitoring extensions
- Watcher, observability, safety, security: incremental improvements

Frontend:
- NativeWorkspace: major overhaul with new pane system
- New panels: BrowserProfiles, Channels, Devices, GoogleWorkspaceAdmin,
  IntegrationQuickstart, IntegrationRouting, PluginSdk, Routing,
  SenderVerification, Webhooks
- IntegrationsPanel: full rewrite with provider management
- API client: expanded endpoint coverage
- Types: comprehensive type definitions for all new features
- Styles: updated theme and layout system

Infrastructure:
- Dockerfile: Ollama CLI, Remotion template, WhatsApp bridge installs
- docker-compose: resource limits, socket proxy tuning
- CI: release workflow updates
- Skills: renamed ACTION.md → SKILL.md across all skills

Cleanup:
- Removed output/video_flicker_check and output/video_frames artifacts
- Removed security_best_practices_report.md
- Added *.db to .gitignore

+++ src/storage/entities/document.rs
@@ -11,7 +11,7 @@ pub struct Model {
     pub content_type: String,
     pub project_id: Option<String>,
     pub chunk_count: i32,
-    pub file_size: i64,
+    pub file_size: i32,
     pub created_at: String,
 }
